Generalize Parser to operate on anything
Makes the `Parser` struct generic on its input and output types, so
it's no longer limited to operate on `String` and
`Substring`. Specific functions and extensions for operating on
`String` and working with `Character` and `CharacterSet` were already
extracted into the `Tokens.swift` file. `Tokens.swift` has been
renamed `StringParsing.swift` and the single useful reference to
`Parser` in an extension defining `run: (String)->(A, Substring)` has
been moved to that file.

A `Parser` type generic on its unput will hopefully be useful in
writing parsers for binary data formats.