Add Double parsing
Follows the Swift language definition of decimal floating point Doubles.
Verification of the structure through parsing is formal. But
construction of the Double value reconstructs the string and passes it
to a Double initializer. I took a more formal approach at first and
found too great a loss of precision. The current method loses precision
as well.

The tests in FFCParserCombinatorTests.testDouble establish some of the
precision limits for Normal, Nonzero and Finite extremes.