Fix Double parsing
Previously the fractional portion of floating point numbers was parsed
as a `UInt`. Unfortunately that mean floating point numbers whose
fractional part had leading zeros were being incorrectly parsed. For
example, `1.001` was incorrectly parsed as `1.1`. This change switches
the parsing of the fractional part of a floating point number to be
intermediated by a string.